ledger: refactor validateValue from CommandsValidator to new ValueValidator
