+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/ledger/api/validation/CommandsValidator.scala
@@ -6,7 +6,6 @@ package com.digitalasset.ledger.api.validation
 import com.digitalasset.api.util.TimestampConversion
 import com.digitalasset.daml.lf.command._
 import com.digitalasset.daml.lf.data._
-import com.digitalasset.daml.lf.value.Value.ValueUnit
 import com.digitalasset.ledger.api.domain
 import com.digitalasset.ledger.api.v1.commands.Command.Command.{
   Create => ProtoCreate,
@@ -16,16 +15,6 @@ import com.digitalasset.ledger.api.v1.commands.Command.Command.{
   ExerciseByKey => ProtoExerciseByKey
 }
 import com.digitalasset.ledger.api.v1.commands.{Command => ProtoCommand, Commands => ProtoCommands}
-import com.digitalasset.ledger.api.v1.value.Value.Sum
-import com.digitalasset.ledger.api.v1.value.{
-  Identifier,
-  RecordField,
-  Value,
-  List => ApiList,
-  Map => ApiMap,
-  Variant => ApiVariant,
-  Enum => ApiEnum
-}
 import com.digitalasset.daml.lf.value.{Value => Lf}
 import com.digitalasset.ledger.api.domain.LedgerId
 import com.digitalasset.platform.common.PlatformTypes.asVersionedValueOrThrow
@@ -38,6 +27,9 @@ import scalaz.syntax.tag._
 import scala.collection.immutable
 
 final class CommandsValidator(ledgerId: LedgerId, identifierResolver: IdentifierResolver) {
+
+  private[this] val valueValidator = new ValueValidator(identifierResolver)
+  import valueValidator._
 
   def validateCommands(commands: ProtoCommands): Either[StatusRuntimeException, domain.Commands] =
     for {
@@ -152,97 +144,5 @@ final class CommandsValidator(ledgerId: LedgerId, identifierResolver: Identifier
       case ProtoEmpty =>
         Left(missingField("command"))
     }
-
-  private def validateRecordFields(recordFields: Seq[RecordField])
-    : Either[StatusRuntimeException, ImmArray[(Option[Ref.Name], domain.Value)]] =
-    recordFields
-      .foldLeft[Either[StatusRuntimeException, BackStack[(Option[Ref.Name], domain.Value)]]](
-        Right(BackStack.empty))((acc, rf) => {
-        for {
-          fields <- acc
-          v <- requirePresence(rf.value, "value")
-          value <- validateValue(v)
-          label <- if (rf.label.isEmpty) Right(None) else requireIdentifier(rf.label).map(Some(_))
-        } yield fields :+ label -> value
-      })
-      .map(_.toImmArray)
-
-  def validateValue(value: Value): Either[StatusRuntimeException, domain.Value] = value.sum match {
-    case Sum.ContractId(cId) =>
-      Ref.ContractIdString
-        .fromString(cId)
-        .left
-        .map(invalidArgument)
-        .map(coid => Lf.ValueContractId(Lf.AbsoluteContractId(coid)))
-    case Sum.Decimal(value) =>
-      Decimal.fromString(value).left.map(invalidArgument).map(Lf.ValueDecimal)
-
-    case Sum.Party(party) =>
-      Ref.Party.fromString(party).left.map(invalidArgument).map(Lf.ValueParty)
-    case Sum.Bool(b) => Right(Lf.ValueBool(b))
-    case Sum.Timestamp(micros) =>
-      Time.Timestamp.fromLong(micros).left.map(invalidArgument).map(Lf.ValueTimestamp)
-    case Sum.Date(days) =>
-      Time.Date.fromDaysSinceEpoch(days).left.map(invalidArgument).map(Lf.ValueDate)
-    case Sum.Text(text) => Right(Lf.ValueText(text))
-    case Sum.Int64(value) => Right(Lf.ValueInt64(value))
-    case Sum.Record(rec) =>
-      for {
-        recId <- validateOptionalIdentifier(rec.recordId)
-        fields <- validateRecordFields(rec.fields)
-      } yield Lf.ValueRecord(recId, fields)
-    case Sum.Variant(ApiVariant(variantId, constructor, value)) =>
-      for {
-        validatedVariantId <- validateOptionalIdentifier(variantId)
-        validatedConstructor <- requireName(constructor, "constructor")
-        v <- requirePresence(value, "value")
-        validatedValue <- validateValue(v)
-      } yield Lf.ValueVariant(validatedVariantId, validatedConstructor, validatedValue)
-    case Sum.Enum(ApiEnum(enumId, value)) =>
-      for {
-        validatedEnumId <- validateOptionalIdentifier(enumId)
-        validatedValue <- requireName(value, "value")
-      } yield Lf.ValueEnum(validatedEnumId, validatedValue)
-    case Sum.List(ApiList(elems)) =>
-      elems
-        .foldLeft[Either[StatusRuntimeException, BackStack[domain.Value]]](Right(BackStack.empty))(
-          (valuesE, v) =>
-            for {
-              values <- valuesE
-              validatedValue <- validateValue(v)
-            } yield values :+ validatedValue)
-        .map(elements => Lf.ValueList(FrontStack(elements.toImmArray)))
-    case _: Sum.Unit => Right(ValueUnit)
-    case Sum.Optional(o) =>
-      o.value.fold[Either[StatusRuntimeException, domain.Value]](Right(Lf.ValueOptional(None)))(
-        validateValue(_).map(v => Lf.ValueOptional(Some(v))))
-    case Sum.Map(m) =>
-      val entries = m.entries
-        .foldLeft[Either[StatusRuntimeException, FrontStack[(String, domain.Value)]]](
-          Right(FrontStack.empty)) {
-          case (acc, ApiMap.Entry(key, value0)) =>
-            for {
-              tail <- acc
-              v <- requirePresence(value0, "value")
-              validatedValue <- validateValue(v)
-            } yield (key -> validatedValue) +: tail
-        }
-
-      for {
-        list <- entries
-        map <- SortedLookupList.fromImmArray(list.toImmArray).left.map(invalidArgument)
-      } yield Lf.ValueMap(map)
-
-    case Sum.Empty => Left(missingField("value"))
-  }
-
-  private def validateOptionalIdentifier(
-      variantIdO: Option[Identifier]): Either[StatusRuntimeException, Option[Ref.Identifier]] = {
-    variantIdO
-      .map { variantId =>
-        identifierResolver.resolveIdentifier(variantId).map(Some.apply)
-      }
-      .getOrElse(Right(None))
-  }
 
 }
