+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/ledger/api/ValueConversionRoundTripTest.scala
@@ -5,7 +5,6 @@ package com.digitalasset.ledger.api
 
 import com.digitalasset.daml.lf.data.Time
 import com.digitalasset.daml.lf.testing.parser.Implicits._
-import com.digitalasset.ledger.api.domain.LedgerId
 import com.digitalasset.ledger.api.v1.value.Value.Sum
 import com.digitalasset.ledger.api.v1.value.{
   List => ApiList,
@@ -13,7 +12,7 @@ import com.digitalasset.ledger.api.v1.value.{
   Optional => ApiOptional,
   _
 }
-import com.digitalasset.ledger.api.validation.{CommandsValidator, ValidatorTestUtils}
+import com.digitalasset.ledger.api.validation.{ValueValidator, ValidatorTestUtils}
 import com.digitalasset.platform.participant.util.LfEngineToApi
 import com.digitalasset.platform.server.api.validation.IdentifierResolver
 import com.google.protobuf.empty.Empty
@@ -54,8 +53,7 @@ class ValueConversionRoundTripTest
          }
          """
 
-  private val commandValidator = new CommandsValidator(
-    LedgerId("ledger-id"),
+  private val commandValidator = new ValueValidator(
     new IdentifierResolver(_ => Future.successful(Some(pkg)))
   )
